+++ Assets/Tiny Rhino/Resources/Prefab/Foods/0.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 4855156344350401786}
   - component: {fileID: 7823455440604920555}
   - component: {fileID: -1599718479371151816}
+  - component: {fileID: 762298985502247295}
   - component: {fileID: 4220769266337680323}
   m_Layer: 0
   m_Name: 0
@@ -94,6 +95,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 538bf980e1d702f4d87860a596a20736, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &762298985502247295
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 543606517614203205}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c165f00eea552b048bbce39130c8ff8a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!61 &4220769266337680323
