+++ Assets/Tiny Rhino/Scripts/ChangeFood.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 538bf980e1d702f4d87860a596a20736
+guid: 720c5454425c58d40aee09999d80ac07
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
