+++ Assets/Tiny Rhino/Scripts/CheckIfClick.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 538bf980e1d702f4d87860a596a20736
+guid: c165f00eea552b048bbce39130c8ff8a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
