+++ Assets/Tiny Rhino/Scripts/CreatePlate.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 538bf980e1d702f4d87860a596a20736
+guid: 1ce6b97d1bb96834d897a2175de49e20
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
